requirements.txt
Pillow==11.1.0
scikit-learn==1.6.1
tensorflow==2.18.0
matplotlib==3.10.0